requirements.txt
Appium-Python-Client==0.50
astroid==2.3.3
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
PyYAML==5.3
selenium==3.141.0
six==1.14.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
